backend/Texel-Art-Media/requirements.txt
absl-py
attrs>=19.1.0
matplotlib
opencv-contrib-python>=4.5.5.64
protobuf==3.20.2
mediapipe
